refactor: move `FooterMode` to `config_toml`
